integration-tests/requirements.txt
pytest~=8.3.4
httpx~=0.27.0
pytest-asyncio~=0.25.1
asyncpg
python-dotenv~=1.0.1
faker~=33.3.0
uuid~=1.30
allure-pytest
allure-python-commons~=2.13.5
bcrypt
pytest-dotenv